+++ CHANGES.rst
@@ -2,7 +2,13 @@ Changelog
 =========
 
 
-1.0a5 (unreleased)
+1.0.1 (unreleased)
+------------------
+
+- Nothing changed yet.
+
+
+1.0.0 (2018-10-04)
 ------------------
 
 - Add check if Plone Site element was indexed.